requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - MongoDB
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1  # MongoDB Python driver

# Image processing and EXIF
Pillow==10.1.0
piexif==1.1.3

# ML (PyTorch - for future model integration)
torch==2.1.1
torchvision==0.16.1

# Testing
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
uuid6==2024.1.12

